Renamed project modules, messages as per spec
