--- automate-teamcity-common/pom.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <artifactId>automate-teamcity</artifactId>
-    <groupId>com.browserstack</groupId>
-    <version>1.0-SNAPSHOT</version>
-  </parent>
-  <artifactId>automate-teamcity-common</artifactId>
-  <packaging>jar</packaging>
-</project>
