+++ browserstack-teamcity-agent/src/main/resources/META-INF/build-agent-plugin-automate-teamcity.xml
