+++ browserstack-teamcity-agent/teamcity-plugin.xml
