+++ browserstack-teamcity-server/src/main/java/com/browserstack/automate/ci/teamcity/AutomateReportBuilder.java
