+++ browserstack-teamcity-server/src/main/resources/META-INF/build-server-plugin-automate-teamcity.xml
