+++ pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.browserstack</groupId>
-  <artifactId>automate-teamcity</artifactId>
+  <artifactId>browserstack-teamcity</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>pom</packaging>
   <properties>
@@ -45,9 +45,9 @@
       </pluginManagement>
   </build>
   <modules>
-    <module>automate-teamcity-server</module>
-    <module>automate-teamcity-agent</module>
-    <module>automate-teamcity-common</module>
+    <module>browserstack-teamcity-server</module>
+    <module>browserstack-teamcity-agent</module>
+    <module>browserstack-teamcity-common</module>
     <module>build</module>
   </modules>
 </project>
